+++ lib/puppet/provider/flatpak_remote/flatpak_remote.rb
@@ -1,7 +1,7 @@
 # Copyright
 # ---------
 #
-# Copyright 2017 Bryan Wyatt, unless otherwise noted.
+# Copyright 2017, 2018 Bryan Wyatt, unless otherwise noted.
 #
 # This file is part of brwyatt-flatpak.
 #
@@ -18,30 +18,96 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with brwyatt-flatpak.  If not, see <http://www.gnu.org/licenses/>.
 
+require File.expand_path('../../../../../../inifile/lib/puppet/util/ini_file', __FILE__)
+
 Puppet::Type.type(:flatpak_remote).provide(:flatpak_remote) do
   commands flatpak: '/usr/bin/flatpak'
+  commands gpg: '/usr/bin/gpg'
+
+  @config_file = Puppet::Util::IniFile.new('/var/lib/flatpak/repo/config', '=')
+
+  def initialize(value={})
+    puts("initialize: #{value}")
+    super(value)
+    @property_flush = {}
+  end
+
+  def self.get_remote(remote)
+    puts('get_remote')
+    section_name = "remote \"#{remote}\""
+
+    if @config_file.section_names.include? section_name then
+      settings = @config_file.get_settings(section_name)
+      {
+        name: remote,
+        ensure: :present,
+        url: settings['url'],
+        gpg_verify: settings['gpg-verify'].to_s.downcase == 'true',
+        gpg_verify_summary: settings['gpg-verify-summary'].to_s.downcase == 'true',
+      }
+    else
+      {
+        name: remote,
+        ensure: :absent,
+      }
+    end
+  end
+
+  def self.instances
+    puts('instances')
+    remotes = []
+
+    @config_file.section_names.each do |section_name|
+      if match = section_name.match(%r{\Aremote "([^"]*)"\Z})
+        remotes << new(get_remote(match.captures[0]))
+      end
+    end
+
+    puts('end::instances')
+    puts("Remotes: #{remotes}")
+    remotes.compact
+  end
+
+  def self.prefetch(resources)
+    puts('prefetch')
+    remotes = instances
+
+    resources.each_key do |name|
+      provider = remotes.find { |remote|
+        remote.name == resources[name].parameters[:name].value
+      }
+
+      if provider then
+        puts("PROVIDER FOUND! #{name}")
+        resources[name].provider = provider
+      end
+    end
+  end
 
   def create
-    args = ['remote-add']
-    if resource[:from]
-      args.push('--from')
-    end
-    args.push(resource[:name], resource[:location])
-    flatpak(args)
+    puts('create')
+    @property_flush[:ensure] = :present
   end
 
   def destroy
-    flatpak 'remote-delete', resource[:name]
+    puts('destroy')
+    @property_flush[:ensure] = :absent
   end
 
   def exists?
-    method(:flatpak).call(['remotes']).split("\n").each do |line|
-      if %r{^#{Regexp.escape(resource[:name])}\s}.match?(line)
-        return true
-      end
-    end
-    false
+    puts("exists: #{@property_flush[:ensure]}")
+    @property_flush[:ensure] == :present
   end
+
+  def flush
+    puts('flush')
+    if @property_flush[:ensure] == :absent
+    end
+
+    @property_hash = self.class.get_remote(resource[:name])
+  end
+
+  mk_resource_methods
 end
 
 # vim: ts=2 sts=2 sw=2 expandtab